+++ packages/design-system/.eslintrc.cjs
@@ -1,0 +1,1 @@
+module.exports = require('../../tooling/.eslintrc.cjs');