+++ packages/design-system/.storybook/preview.ts
@@ -1,0 +1,32 @@
+import type { Preview } from '@storybook/react';
+import { CssBaseline, ThemeProvider } from '@mui/material';
+import { createTheme } from '@mui/material/styles';
+import React from 'react';
+
+import { muiThemeJson } from '../src';
+import '../src/theme/tokens.css';
+import '../src/theme/theme.css';
+
+const theme = createTheme(muiThemeJson);
+
+const preview: Preview = {
+  decorators: [
+    (Story) => (
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <Story />
+      </ThemeProvider>
+    )
+  ],
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/
+      }
+    }
+  }
+};
+
+export default preview;