+++ src/index.d.ts
@@ -1,6 +1,0 @@
-export declare const muiThemeJson: any;
-export declare const themeFlat: any;
-export declare const coreFlat: any;
-export declare const themeMuiFlat: any;
-export declare const coreMuiFlat: any;
-